services/news-signal/pyproject.toml
[project]
name = "news-signal"
version = "0.1.0"
description = "MLOps Crypto Trading Real-time Data Service"
keywords = ["MLOps", "Trading", "LLM", "Machine Learning"]
authors = [{ name = "Benito Martin"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.12.5",
    "llama-index-llms-anthropic>=0.5.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.10.3",
    "llama-index-llms-ollama>=0.5.0",
    "ollama>=0.4.4",
    "pre-commit>=4.0.1",
    "quixstreams>=3.4.0",
    "fire>=0.7.0",
    "pandas>=2.2.3",
]

[project.optional-dependencies]
dev = [
    "pip>=24.3.1",
    "ruff>=0.8.3",
]

# Build system configuration
[build-system]
requires = ["hatchling>=1.23.0"]
build-backend = "hatchling.build"

# # Hatch configuration
[tool.hatch.build]
packages = ["llms"]
